# Performance enhancements to collecting labeled metrics