--- a/prometheus-net/Advanced/Collector.cs
+++ b/prometheus-net/Advanced/Collector.cs
@@ -39,15 +39,19 @@ namespace Prometheus.Advanced
 		}
 
 		private T GetOrAddLabelled(LabelValues key)
-        {
-            return _labelledMetrics.GetOrAdd(key, labels1 =>
-            {
-                var child = new T();
-                child.Init(this, labels1);
-                return child;
-            });
-        }
+		{
+		    T val;
+		    if (_labelledMetrics.TryGetValue(key, out val))
+		        return val;
 
+		    val = new T();
+		    val.Init(this, key);
+		    _labelledMetrics.TryAdd(key, val);
+		    key.InitWireLabels();
+		    
+		    return val;
+        }
+        
         protected T Unlabelled
         {
             get { return _unlabelledLazy.Value; }