- Replaced the use of `ConcurrentDictionary<T, K>.GetOrAdd` in `Collector.GetOrAddLabelled` as the function was allocating a closure
- Made `LabelValues` a struct to avoid heap allocations
- `LabelValues` precomputes it's hash code now- as this never changes, it doesn't make sense to re-compute it. Saves CPU when repeatedly recording metrics with the same labels.
- `LabelValues.LabelValues` is only initialised once per unique combination of label values, this saves CPU + memory when repeatedly recording metrics with the same labels.

Changes allow recording metrics with labels to be ~3x faster and allocate ~12x less memory

|       Method |     Mean |     Error |    StdDev |  Gen 0 | Allocated |
------------------- |---------  |---------- |---------- |------- |---------- |
 RecordLabelsNew | 321.5 ns | 9.348 ns | 13.99 ns | 0.0176 |      56 B |
 RecordLabelsOld | 1.169 us | 0.0204 us | 0.0306 us | 0.1888 |     600 B |