@phnx47 2017-11-22
@djluck Can you attach benchmark code? I think, I can use it in my fork: https://github.com/phnx47/Prometheus.Client