@djluck 2017-11-22
@phnx47 sure, it is:

```
void Main()
{
	var config = DefaultConfig.Instance
		.With(new Job("X64-Clr-RyuJit", RunMode.Medium, EnvMode.Clr)
		{
			Accuracy = { RemoveOutliers = false },
			Env = { Runtime = Runtime.Clr, Platform = Platform.X64, Jit = Jit.RyuJit },
			
		})
		.With(MemoryDiagnoser.Default);

	BenchmarkRunner
		.Run<PrometheusBenchmark>(config);
}

public class PrometheusBenchmark
{
	private Counter nolabels, labels;
	public PrometheusBenchmark()
	{
		nolabels = Prometheus.Metrics.CreateCounter("my_counter", "help");
		labels = Prometheus.Metrics.CreateCounter("my_counter_labels", "help", "label_one", "label_two", "label_three", "label_four");
	}
	
//	[Benchmark]
//	public void RecordNoLabels()
//	{
//		nolabels.Inc();
//	}

	[Benchmark]
	public void RecordLabels()
	{
		labels.Labels("label_one_val", "label_two_val", "label_three_val", "label_four_val").Inc();
	}
}
```