@phnx47 2017-11-22
prometheus-net is slowly developing, I can't waiting. I fixed many bug, supported netstandard and tested it. I'm working on several extensions for my projects and publish on GitHub