@phnx47 2017-11-23
> Replaced the use of ConcurrentDictionary<T, K>.GetOrAdd in Collector.GetOrAddLabelled as the function was allocating a closure

It isn't correct. Tests no working, when you combination with Label and without Label https://github.com/andrasm/prometheus-net/blob/77428872f1ae94f81f72badf67eaa58654bce6a9/prometheus-net.tests/MetricsTests.cs#L48