@sandersaares 2018-01-25
Thank you for the contribution!

I have manually merged an adjusted variant of this PR into the v2 branch to avoid merge conflicts due to some structural changes in the codebase.

I am seeing (with slightly different benchmark code) approximately similar gains as you did.